+++ packages/cli/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Multi-chain project initialization error
+
 ## [5.3.0] - 2024-10-21
 ### Changed
 - Improve codegen error messages (#2567)